+++ src/main/java/com/reason/lang/ocaml/OclLanguage.java
@@ -6,7 +6,7 @@ public class OclLanguage extends Language {
     public static final OclLanguage INSTANCE = new OclLanguage();
 
     private OclLanguage() {
-        super("Ocaml");
+        super("OCaml");
     }
 
 }